ì오류 수정

+++ src/main/java/com/aboutme/springwebservice/board/repository/QnACategoryRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 public interface QnACategoryRepository extends JpaRepository<QnACategory, Long> {
 
     @Query(value = "DELETE FROM QnA_Category WHERE seq = ?1", nativeQuery = true)
-    void delCardQuestion(int categorySeq);
+    void delCardQuestion(long categorySeq);
 
     QnACategory findBySeq(long seq);
